+++ spindle/src/test/java/com/salesforce/ouroboros/spindle/TestAppender.java
@@ -245,7 +245,7 @@ public class TestAppender {
             }
 
             @Override
-            public void registerReplicator(long id, Replicator replicator) {
+            public void registerReplicator(int id, Replicator replicator) {
                 throw new UnsupportedOperationException();
             }
         };
